reverted the change, int instead of long should be used for Android plugin

+++ plugin/PlayAssetDelivery/src/main/java/com/google/play/core/godot/assetpacks/PlayAssetDelivery.java
@@ -199,7 +199,7 @@ public class PlayAssetDelivery extends GodotPlugin {
    * @param packNamesArray String Array for all the packs to be fetched
    * @param signalID identifier used to track mapping of signals to Tasks
    */
-  public void fetch(String[] packNamesArray, long signalID) {
+  public void fetch(String[] packNamesArray, int signalID) {
     List<String> packNames = Arrays.asList(packNamesArray);
 
     OnSuccessListener<AssetPackStates> fetchSuccessListener =
@@ -252,7 +252,7 @@ public class PlayAssetDelivery extends GodotPlugin {
    * @param packNamesArray String Array for all the packs to request states
    * @param signalID identifier used to track mapping of signals to Tasks
    */
-  public void getPackStates(String[] packNamesArray, long signalID) {
+  public void getPackStates(String[] packNamesArray, int signalID) {
     List<String> packNames = Arrays.asList(packNamesArray);
 
     OnSuccessListener<AssetPackStates> getPackStatesSuccessListener =
@@ -281,7 +281,7 @@ public class PlayAssetDelivery extends GodotPlugin {
    * @param packName name of the asset pack to be removed
    * @param signalID identifier used to track mapping of signals to Tasks
    */
-  public void removePack(String packName, long signalID) {
+  public void removePack(String packName, int signalID) {
     OnSuccessListener<Void> removePackOnSuccessListener =
         result -> emitSignalWrapper(REMOVE_PACK_SUCCESS, signalID);
     OnFailureListener removePackOnFailureListener =
@@ -305,7 +305,7 @@ public class PlayAssetDelivery extends GodotPlugin {
    *
    * @param signalID identifier used to track mapping of signals to Tasks
    */
-  public void showCellularDataConfirmation(long signalID) {
+  public void showCellularDataConfirmation(int signalID) {
     OnSuccessListener<Integer> showCellularDataConfirmationSuccessListener =
         result -> emitSignalWrapper(SHOW_CELLULAR_DATA_CONFIRMATION_SUCCESS, result, signalID);
     OnFailureListener showCellularDataConfirmationFailureListener =
